requirements.txt
pandas
python-dotenv
supabase
postgrest
matplotlib